rename nbgrader module to setup_course
