+++ src/illumidesk/setup_course/app.py
@@ -12,10 +12,10 @@ from io import StringIO
 from pathlib import Path
 from secrets import token_hex
 
-import docker
 from quart import Quart
 from quart import request
 from .course import Course
+from .utils import SetupUtils
 
 
 logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
@@ -92,30 +92,3 @@ def update_jupyterhub_config(course: Course):
     with lock:
         with jupyterhub_config_json.open('r+') as config:
             json.dump(cache, config)
-
-
-
-class SetupUtils:
-    """
-    Utils class used to manage course setup configurations and updates.
-    """
-    def __init__(self):
-        self.docker_client = docker.from_env()
-
-    def restart_jupyterhub(self):
-        """
-        Restart jupyterhub using the docker client after updating configs.
-
-        Raises:
-          ContainerException if the container could not be restarted.
-        """
-        logger.debug('Received request to restart JupyterHub')
-        jupyterhub_container_name = os.environ.get('JUPYTERHUB_SERVICE_NAME')
-        containers = self.docker_client.containers.list(filters={'name': f'{jupyterhub_container_name}'})
-        for container in containers:
-            logger.debug('Found container %s, restarting ...' % jupyterhub_container_name)
-            try:
-                container.restart()
-            except docker.errors.NotFound:
-                logger.error('Grader container not found')
-
